requirements.txt
fastapi
uvicorn
python-dotenv
groq
google-generativeai
cohere
duckduckgo-search
httpx
beautifulsoup4
